rename post article dialog
