+++ app/src/main/res/layout/dialog_post_article.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="viewModel"
-            type="com.willy.metu.dialog.ArticleDialogViewModel" />
+            type="com.willy.metu.postarticle.PostArticleDialogViewModel" />
 
     </data>
